Enable permission management when broadcaster is initialized
Prior to the upgrade to SDK 0.9.12, the affected permission request would have fired
a bit early in some apps, at plugin initialization. Instead, we relied on
the other permission request when starting the first broadcast.

After the upgrade to SDK 0.9.12, the affected permission request fires
when applicationId is set. This makes the viewfinder usable prior to
starting a broadast, which is often nice.

If desired, permission request options could be provided via the JavaScript API
in the future.