Queue commands to avoid race conditions
If the application code has independent triggers,
for example navigation between two pages in an Ionic app,
it is quite easy to trigger a second command before
the callback of the first command has fired.

As of this commit, this is avoided on the plugin level,
by adding all commands to a Promise-based queue.

Theoretical drawback: could delay commands undesiredly.