Cast parent reference to ViewGroup base class instead of more specialized view types
Previously with Cordova - including in the current Travis-CI tests
with cordova-paramedic, FrameLayout seems to be the actual class used.

Capacitor on the other hand, seems to use CoordinatorLayout, leading
to a runtime failure:
java.lang.ClassCastException: android.support.design.widget.CoordinatorLayout cannot be cast to android.widget.FrameLayout

Similarly, some versions or configurations of cordova-android
seem to use LinearLayout, leading to a similar failure:
https://github.com/bambuser/cordova-plugin-bambuser/issues/5

Using webView.getView().getParent() directly does not work, nor does
casting explicitly to View: addView() is not defined in those cases,
whereas ViewGroup seems to be an intermediate base class for all of the
classes mentioned above and seems to implement addView().
https://developer.android.com/reference/android/view/ViewGroup.html